requirements.txt
Flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
Pillow==10.3.0
tensorflow==2.15.0
keras==2.15.0